sistema-ortomedica/backend/requirements.txt
playwright==1.47.0
python-dotenv==1.0.1
mysql-connector-python==9.0.0
openpyxl==3.1.2
xlrd==1.2.0

# API
fastapi==0.104.1
uvicorn[standard]==0.23.2
python-jose[cryptography]==3.3.0

# Auth hashing
passlib[bcrypt]==1.7.4
# IMPORTANT: bcrypt>=4.1 breaks passlib's runtime checks; pin to 4.0.1
bcrypt==4.0.1

pytz==2023.3

# Scraping/parsing deps
beautifulsoup4==4.12.3
lxml==4.9.3